Adding UserPageController.tsx to add the ChatBox to UserPage, and moved the Controllers to separate folder called Controlers in Frontend.
